refactor main function to own file
