+++ src/fallback-functions/adSense.js
