+++ src/fallback-functions/createImage.js
