+++ src/fallback-functions/createPostMessageScript.js
