+++ src/fallback-functions/flashTalking.js
